Refactor codebase
Refactored various components and hooks in the codebase for improved readability and maintainability.
[skip gpt_engineer]
