+++ src/components/chat/MessageContent.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import ReactMarkdown from 'react-markdown';
 import DOMPurify from 'dompurify';
-import Linkify from 'react-linkify';
-import { CitationSource } from './CitationSource';
+import { MarkdownContent } from './MarkdownContent';
+import { formatContentWithSources } from '@/utils/messageFormatting';
 
 interface MessageContentProps {
   content: string;
@@ -26,32 +25,7 @@ export const MessageContent = ({
   selectedCitation 
 }: MessageContentProps) => {
   const sanitizedContent = DOMPurify.sanitize(content);
-  
-  const formatContentWithSources = (content: string) => {
-    if (!sources) return content;
-    
-    let formattedContent = content;
-    
-    // Remove the "Sources utilisées :" section if it exists
-    formattedContent = formattedContent.split(/Sources utilisées :/).shift() || formattedContent;
-    
-    // Remove any [object Object] artifacts
-    formattedContent = formattedContent.replace(/,\s*\[object Object\]/g, '');
-    
-    // Clean up any trailing commas
-    formattedContent = formattedContent.replace(/,\s*$/g, '');
-    
-    // Remove any empty source brackets
-    formattedContent = formattedContent.replace(/\[\d+\]\s*,/g, '[$1]');
-    
-    return formattedContent.trim();
-  };
-
-  const componentDecorator = (href: string, text: string, key: number) => (
-    <a href={href} key={key} target="_blank" rel="noopener noreferrer" className="text-blue-600 hover:text-blue-800">
-      {text}
-    </a>
-  );
+  const formattedContent = formatContentWithSources(sanitizedContent, sources);
 
   return (
     <div className="space-y-4">
@@ -70,34 +44,7 @@ export const MessageContent = ({
         return null;
       })}
       
-      <div className="prose prose-sm max-w-none">
-        <Linkify componentDecorator={componentDecorator}>
-          <ReactMarkdown components={{
-            a: ({ node, ...props }) => (
-              <a {...props} target="_blank" rel="noopener noreferrer" className="text-blue-600 hover:text-blue-800" />
-            ),
-            li: ({ node, ...props }) => (
-              <li {...props} className="my-1" />
-            ),
-            ul: ({ node, ...props }) => (
-              <ul {...props} className="list-disc pl-4 my-2 space-y-1" />
-            ),
-            ol: ({ node, ...props }) => (
-              <ol {...props} className="list-decimal pl-4 my-2 space-y-1" />
-            ),
-            p: ({ node, ...props }) => (
-              <p {...props} className="my-2 leading-relaxed" />
-            ),
-            code: ({ node, className, children, ...props }) => (
-              className ? 
-                <code {...props} className="bg-gray-100 rounded px-1 py-0.5">{children}</code> :
-                <code {...props} className="block bg-gray-800 text-gray-100 p-4 rounded-lg overflow-x-auto">{children}</code>
-            )
-          }}>
-            {formatContentWithSources(sanitizedContent)}
-          </ReactMarkdown>
-        </Linkify>
-      </div>
+      <MarkdownContent content={formattedContent} />
 
       {sources && sources.length > 0 && (
         <div className="mt-4">
